move and rename FT for freeform modules, add the other freeform Ansible modules for testing
